pyproject.toml
[tool.poetry]
name = "BetterStatcord.py"
version = "2.0.1"
description = "A better Statcord API library for Python + Discord.py"
authors = ["Milo Weinberg <iapetus011@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Iapetus-11/betterstatcord.py"
packages = [
  { include = "statcord" }
]

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.7.4"
psutil = "^5.8.0"
"discord.py" = "^1.7.1"

[tool.poetry.dev-dependencies]
black = "^21.4b0"
flake8 = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
